windows: Made GUID type available to other OS's
