+++ windows/types.go
@@ -1,0 +1,8 @@
+package windows
+
+type GUID struct {
+	Data1 uint32
+	Data2 uint16
+	Data3 uint16
+	Data4 [8]byte
+}
